requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
sqlite3
python-multipart==0.0.6
python-dotenv==1.0.0

# OpenAI and AI services
openai==1.3.7
anthropic==0.7.8

# Voice processing
elevenlabs==0.2.26
openai-whisper==20231117
pyaudio==0.2.11
pydub==0.25.1
ffmpeg-python==0.2.0

# RAG and vector database
chromadb==0.4.18
sentence-transformers==2.2.2
langchain==0.0.350
tiktoken==0.5.2

# WhatsApp and messaging
twilio==8.10.0
requests==2.31.0
flask==3.0.0

# Web interface
websockets==12.0
jinja2==3.1.2

# Utilities
pandas==2.1.4
numpy==1.26.2
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# WebSocket support
websockets==12.0

# Speechmatics integration
speechmatics-python==0.6.0

# Development
black==23.11.0
flake8==6.1.0
pre-commit==3.6.0